0.0.2: API Manager, User Manager, etc.
New button group for the page buttons inside the PopupPage

Updated the font to make it work when scaled as before it devolved into pixel soup when viewed from far away.

New styleboxes for the new CustomButton scene.

Made a test 'SettingsPage' scene to make sure the PopupPage page-switching worked, it will be actually populated later.

New APIManager autoload which will handle all the API stuff, for now you can just about download an exe for NPS from github, have not tested any of the others.

Two new colours in 'highlight 1' and 'highlight 2' for the ColourManager, and functions to get and apply the colour settings to and from a dictionary.

New UserManager which will handle saving and loading user data, for now that is the colour settings and an unused default download location.

PopupPage now has settable data which after it is set then populates all the tab buttons and page scenes.