0.0.6: Key Validation, Fixes, etc.
Removed many unused icons.

Added new Socials icon for the new Socials tab which shows links to some of the NZG platforms, each has its own company logo.

Made the export pre-sets.

New CustomLineEdit scene which is used to set the Itch.io API key in the settings.

New API, Display and General settings pages, and renamed the DetailsPage scene to SettingsDetailsPage to keep with the naming scheme of the rest.

The APIManager now has key validation and checks / error-handling for ensuring that a valid key is present when fetching info and giving a pop-up when it isnt. There are other smaller updates to the APIManager including bug-fixes and rate-limit handling.

New 'ShowAppBackground' settings since the background can sometimes make buttons harder to see / read and 'CloseOnAppLaunch' to close the launcher when launching a programme.

New 'WebSource' button in the app page to show the list of sources with buttons to link to them.

Categories are now listed inside an app's page.

Other smaller changes and fixes.