1.0.3: Better Download Screen, Bugs Fixed, etc.
Removed moving the executable, this may be remade in the future but for now it is gone.

ProductListItem's title now autowraps.

Reinstalling an application now works again.

Fixed the download screen not closing when an error occurs and makes it throw an alert, changed the InfoForNerds to remove the packet usage as it refused to work despite being a simple percentage.